% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNA.R
\name{classifyNA}
\alias{classifyNA}
\alias{classifyRNA}
\alias{classifyDNA}
\title{Classify RNA or DNA structures}
\usage{
classifyRNA(pdbID, length = 3, force = FALSE, ...)

classifyDNA(pdbID, force = FALSE, ...)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure ID in the
Protein Data Bank.}

\item{length}{A positive integer. Minimum numer of nucleotides to consider
RNA as a polymer. An RNA shorter than this threshold is classified in
the NoRNA group.}

\item{force}{A logical to force the query instead of getting presaved data.}

\item{...}{Arguments to be passed to query function (see ?queryFunctions).}
}
\value{
A string with the type of RNA.
}
\description{
\code{classifyRNA} classifes a structure in the following groups:\cr
\itemize{
\item {NoRNA}: the structure does not contain RNA or it is shorter
than a threshold set by "length".\cr
\item {nakedRNA}: the only molecule(s) in the structure is RNA. It can
include ionic ligands, but not organic ligands.\cr
\item {protRNA}: the PDB contains a protein-RNA complex.\cr
\item {DprotRNA}: the PDB contains a protein-RNA complex and the
protein has D aminoacids.\cr
\item {DNARNA}: the PDB contains DNA-RNA.\cr
\item {PNARNA}: the PDB contains PNA-RNA.\cr
\item {ANARNA}: the PDB contains ANA-RNA.\cr
\item {LNARNA}: the PDB contains LNA-RNA.\cr
\item {ligandRNA}: the RNA is interacting with an organic
ligand. Ions are not considered as ligands in this class.\cr
}
}
\details{
\code{classifyDNA} classifes a structure in the following groups:\cr
\itemize{
\item {NoDNA}: the structure does not contain DNA.\cr
\item {nakedDNA}: the only molecule(s) in the structure is DNA. It can
include ionic ligands, but not organic ligands.\cr
\item {protDNA}: the PDB contains a protein-DNA complex.\cr
\item {DprotDNA}: the PDB contains a protein-DNA complex and the
protein has D aminoacids.\cr
\item {DNARNA}: the PDB contains DNA-RNA.\cr
\item {PNADNA}: the PDB contains PNA-DNA.\cr
\item {ligandDNA}: the DNA is interacting with an organic
ligand. Ions are not considered as ligands in this class.\cr
}

In \code{classifyRNA}, nucleic acid hybrids are considered RNA, while in
in \code{classifyDNA} they are considered DNA (e.g. pdb ID 2HVR).
}
\examples{
classifyRNA("1S72")

}
\author{
Diego Gallego
}
