% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEntities.R
\name{countEntities}
\alias{countEntities}
\title{Count entities}
\usage{
countEntities(pdbID, force = FALSE, ...)
}
\arguments{
\item{pdbID}{A 4-character string that matches a structure ID in the
Protein Data Bank.}

\item{force}{A logical to force the query instead of getting presaved data.}

\item{...}{Arguments to be passed to query function (see ?queryFunctions).}
}
\value{
A list with the number of instances of each entity.
}
\description{
For a given pdbID, the function gets the Entity data and counts the number
of instances of the different entities (e.g. the number of different RNA,
the number of different proteins...).
}
\examples{
countEntities("1S72")

}
\author{
Diego Gallego
}
