% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-entities.R
\docType{data}
\name{entities}
\alias{entities}
\title{List of NA containing PDB IDs and related data}
\format{An object of class data.frame
\describe{
\item{pdbID:}{PDB ID.}
\item{RNA:}{Number of different RNA entities.}
\item{DNA:}{Number of different DNA entities.}
\item{Hybrid:}{Number of different Hybrid DNA/RNA entities.}
\item{PNA:}{Number of different PNA entities.}
\item{Prot:}{Number of different Protein (L) entities.}
\item{Dprot:}{Number of different Protein (D) entities.}
\item{Ligands:}{Number of different ligands.}
\item{Water:}{It's 1 when there's water in the structure
and 0 when it is not.}
\item{Other:}{Number of different "other" entities.}
}}
\usage{
data(entities)
}
\value{
data.frame with a list and features of NA PDB IDs.
}
\description{
Number of each type of entity in a set of Nucleic Acid containing
structures.
}
\keyword{datasets}
