% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAltRepres.R
\name{getAltRepres}
\alias{getAltRepres}
\title{Get Alternative Representants}
\usage{
getAltRepres(
  rnalist,
  technique = NULL,
  resol = NULL,
  type = NULL,
  length = 3,
  progressbar = TRUE,
  verbose = FALSE,
  as.df = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{rnalist}{The output of getRNAList.}

\item{technique}{One or more techniques of interest (For correct use, see
example below). For the list of techniques, see "veriNA3d:::.allowedtechs".}

\item{resol}{A positive real number to specify a desired resolution.}

\item{type}{A string indicating the type of desired RNA, according with
the {classifyRNA} function.}

\item{length}{To be passed to {classifyRNA}.}

\item{progressbar}{A logical to print in screen a progress bar.}

\item{verbose}{A logical to print details of the process.}

\item{as.df}{A logical to return the output as a data.frame.}

\item{na.rm}{A logical to clean missing values out of the output.}
}
\value{
A data.frame with info about all the "Equivalence Classes" and
the selected Representants according with the specified conditions.
}
\description{
This function is closely related with getRNAList(). From its
output, the family members of each equivalence class are checked for
desired features. The first member that matches all the desired features
is returned.
}
\examples{
    #rnalist <- getRNAList(release=3.2, threshold="1.5A")
    #alternative <- getAltRepres(rnalist=rnalist, 
    #                            type="nakedRNA")

}
\author{
Diego Gallego
}
