% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCircularDistribution}
\alias{plotCircularDistribution}
\title{Plot a scatter&frequency circular plot for angular data}
\usage{
plotCircularDistribution(
  data,
  clockwise = FALSE,
  start.degree = 0,
  main = NULL
)
}
\arguments{
\item{data}{A numeric vector with the data to plot.}

\item{clockwise}{A logical indicating the sense in which the data should be
displayed.}

\item{start.degree}{An integer with the position in which the data starts
being ploted.}

\item{main}{A string with the title of the plot.}
}
\value{
A circular plot with the input data.
}
\description{
For a vector of angular data (0 to 360), the function plots the
distribution in a circular format.
}
\examples{
    ntinfo <- pipeNucData("1bau")
    C3endo_ntID <- cleanByPucker(ntinfo, pucker="C3'endo")
    C3endo <- ntinfo[ntinfo$ntID \%in\% C3endo_ntID, ]
    plotCircularDistribution(C3endo[, "delta"])

}
\author{
Diego Gallego
}
