% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\name{RMSD}
\alias{RMSD}
\alias{RMSD,CIF,CIF-method}
\alias{RMSD,ANY,ANY-method}
\title{Compute the RMSD between two structures}
\usage{
RMSD(cif1, cif2, sel1 = NULL, sel2 = NULL, ...)

\S4method{RMSD}{CIF,CIF}(cif1 = NULL, cif2 = NULL, sel1 = NULL, sel2 = NULL, ...)

\S4method{RMSD}{ANY,ANY}(cif1 = NULL, cif2 = NULL, sel1 = NULL, sel2 = NULL, ...)
}
\arguments{
\item{cif1}{A CIF or pdb object as otained from
\code{\link[=cifParser]{cifParser()}}, \code{\link[=cifAsPDB]{cifAsPDB()}} or read.cif/read.pdb
(from bio3d package).}

\item{cif2}{A CIF or pdb object as otained from
\code{\link[=cifParser]{cifParser()}}, \code{\link[=cifAsPDB]{cifAsPDB()}} or read.cif/read.pdb
(from bio3d package).}

\item{sel1}{A sel object as obtained from atom.select (from bio3d package).}

\item{sel2}{A sel object as obtained from atom.select (from bio3d package).}

\item{...}{Arguments to be passed to atom.select if desired (only used in
case that sel1 and sel2 are NULL.}
}
\value{
A numeric with the RMSD between the two structures.
}
\description{
Compute the RMSD between two structures
}
\examples{
    cif <- cifParser("2d18")
    model1 <- selectModel(cif=cif, model=1)
    model3 <- selectModel(cif=cif, model=3)
    RMSD(cif1=model1, cif2=model3)

}
\author{
Diego Gallego
}
