% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimByID.R
\name{trimByID}
\alias{trimByID}
\title{Calls trimSphere to generate smaller pdb files}
\usage{
trimByID(
  cif = NULL,
  ntID,
  ntinfo,
  prev = 2,
  post = 2,
  verbose = TRUE,
  file = NULL,
  justbackbone = FALSE,
  ...
)
}
\arguments{
\item{cif}{A cif/pdb object obtained from cifParser/read.pdb respectively
or a pdb ID so that the function can download the data. If NULL, the
function will extract the pdb ID from the ntinfo data frame
(pdbID col).}

\item{ntID}{An integer/string with the desired nucleotide ID for
analysis.}

\item{ntinfo}{a data.frame with the data. It should contain at least the
columns "pdbID", "chain", "model", "resno", "insert" and "ntID" (as the
output of pipeNucData function).}

\item{prev}{Number of desired 5' neigbours to be returned.}

\item{post}{Number of desired 3' neigbours to be returned.}

\item{verbose}{A logical to print details of the process.}

\item{file}{A string to save the output in a pdb formated file. If NULL the
functions just returns the pdb object.}

\item{justbackbone}{A logical to keep only the bakcbone of the output pdb.}

\item{...}{Arguments to be passed to trimSphere (type ?trimSphere for
details).}
}
\value{
A smaller pdb object or a pdb file.
}
\description{
Given a data frame with nucleotide info (as obtained from pipeNucData)
and the desired nucleotide index (ntID), the function returns a pdb
object or file allowing the user to select a number of 5' and 3' neighbors
in sequence and non-conected residues in a cutoff radius.
}
\examples{
    cif <- cifParser("1bau")
    ntinfo <- pipeNucData(cif, torsionals=NULL, distances=NULL, angles=NULL)

    ## Obtain a smaller pdb of the 4th nucleotide +-2 neigbours and a 
    ## sorrounding sphere of 5 Angstroms
    pdb <- trimByID(cif=cif, ntinfo=ntinfo, ntID=4, prev=2, post=2, 
                            cutoff=5)

    ## Same process saving the output in a file:
    trimByID(cif=cif, ntinfo=ntinfo, ntID=4, prev=2, post=2, 
                            cutoff=5, file="output.pdb")

}
\author{
Diego Gallego
}
